+++ doc/source/changelog.rst
@@ -23,6 +23,7 @@ master: 2.3.0
 * Fix lookup of templates of function with template args
 * Fix regression for accessing `char16_t` data member arrays
 * Add custom `__reshape__` method to CPPInstance to allow array cast
+* Prioritize callee exceptions over bindings exceptions
 
 
 2021-11-14: 2.2.0